add information button + change selected contract
